// packages/auth0/src/types.ts
export interface ITagNames {
  readonly auth: string;
  readonly authLogout: string;
  readonly authSession: string;
}

export interface IWritableTagNames {
  auth?: string;
  authLogout?: string;
  authSession?: string;
}

export interface IConfig {
  readonly autoTrigger: boolean;
  readonly triggerAttribute: string;
  readonly tagNames: ITagNames;
}

export interface IWritableConfig {
  autoTrigger?: boolean;
  triggerAttribute?: string;
  tagNames?: IWritableTagNames;
}